basic-app/requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
asgiref==3.9.1
attrs==25.3.0
blinker==1.9.0
Bootstrap-Flask==2.5.0
cachetools==6.1.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.2.1
distro==1.9.0
Flask==3.1.1
gitdb==4.0.12
GitPython==3.1.44
gunicorn==23.0.0
h11==0.16.0
htmltools==0.6.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
narwhals==1.46.0
numpy==2.3.1
openai==1.93.3
orjson==3.10.18
packaging==25.0
pandas==2.3.1
pillow==11.3.0
praw==7.8.1
prawcore==2.4.0
prompt_toolkit==3.0.51
protobuf==6.31.1
pyarrow==20.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytz==2025.2
questionary==2.1.0
referencing==0.36.2
requests==2.32.4
rpds-py==0.26.0
setuptools==80.9.0
shiny==1.4.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.47.1
streamlit==1.46.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
uc-micro-py==1.0.3
update-checker==0.18.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
WTForms==3.2.1
